refactor: move MCP infrastructure code to config package
