+++ gateways/gateway-mcp/src/main/java/io/camunda/gateway/mcp/config/CamundaMcpServerToolSpecificationsAutoConfiguration.java
@@ -8,7 +8,7 @@
 package io.camunda.gateway.mcp.config;
 
 import io.camunda.gateway.mcp.ConditionalOnMcpGatewayEnabled;
-import io.camunda.gateway.mcp.provider.CamundaSyncStatelessMcpToolProvider;
+import io.camunda.gateway.mcp.config.CamundaSyncStatelessMcpToolProvider;
 import io.modelcontextprotocol.server.McpStatelessServerFeatures;
 import io.modelcontextprotocol.server.McpStatelessServerFeatures.SyncToolSpecification;
 import java.util.List;
