+++ gateways/gateway-mcp/src/test/java/io/camunda/gateway/mcp/config/McpRequestBodyValidationTest.java
@@ -1,8 +1,8 @@
-package io.camunda.gateway.mcp.provider;
+package io.camunda.gateway.mcp.config;
 
 import static org.assertj.core.api.Assertions.*;
 
-import io.camunda.gateway.mcp.annotation.McpRequestBody;
+import io.camunda.gateway.mcp.config.McpRequestBody;
 import io.camunda.gateway.mcp.tool.demo.CreateTaskRequest;
 import io.modelcontextprotocol.spec.McpSchema.CallToolResult;
 import jakarta.validation.Valid;
